requirements.txt
llama-cpp-python
python-telegram-bot==20.7
